refactor(saas): moved keypairoom functions to own folder
